apps/backend_django/requirements.txt
Django==5.2
psycopg2-binary==2.9.10
sqlparse==0.5.5
tzdata==2025.3
djangorestframework>=3.14